+++ src/gateway/src/db/entities.rs
@@ -4,7 +4,7 @@ pub struct CreatePayment {
     pub payment_id: Uuid,
     pub full_name: String,
     pub email: String,
-    pub amount: u64,
+    pub amount: u32,
     pub security_question: String,
     pub security_answer: String,
 }
